fix: share canonical TLS adapter types
